+++ applications/fishing-map/src/features/dataviews/dataviews.utils.ts
@@ -1,6 +1,8 @@
+import { kebabCase } from 'lodash'
 import {
   ColorCyclingType,
   Dataset,
+  Dataview,
   DataviewCategory,
   DataviewDatasetConfig,
   DataviewInstance,
@@ -139,4 +141,24 @@ export const getContextDataviewInstance = (
     ],
   }
   return contextDataviewInstance
+}
+
+export const getDataviewInstanceFromDataview = (dataview: Dataview) => {
+  return {
+    id: `${kebabCase(dataview.name)}-${Date.now()}`,
+    dataviewId: dataview.id,
+  }
+}
+
+export const getActivityDataviewInstanceFromDataview = (
+  dataview?: Dataview
+): DataviewInstance<Generators.Type> | undefined => {
+  if (!dataview) return
+  const instance = getDataviewInstanceFromDataview(dataview)
+  return {
+    ...instance,
+    config: {
+      colorCyclingType: 'fill' as ColorCyclingType,
+    },
+  }
 }
